// src/teacher/model/interface/teacher.interface.ts
export interface ITeacher {
  PRO_ID: number;
  PRO_NOME: string;
  PRO_FONE: string;
  PRO_DOCUMENTO?: string;
  PRO_AVATAR?: string;
  PRO_EMAIL: string;
  PRO_ATIVO: boolean;
  PRO_DT_CRIACAO: Date;
  PRO_DT_ATUALIZACAO: Date;
  PRO_DT_NASC: Date;
  PRO_CIDADE: string;
  PRO_ENDERECO: string;
  PRO_NUMERO: string;
  PRO_COMPLEMENTO: string;
  PRO_BAIRRO: string;
  PRO_CEP: string;
}